Put handleSkillSelection into its own file
